# api/app/openapi.yml
openapi: 3.1.0
info:
  title: MADB API
  version: 0.1.0
paths:
...
  /v1/projects/{project_id}/milestones:
    get:
      tags:
      - projects
      summary: Find Milestones By Slug Or Name
      description: 'Resolve milestones for a project by slug or name.


        Returns JSON: {"ok": true, "milestones": [ {id, slug, name, start_date, due_date,
        url}, ... ]}'
      operationId: find_milestones_by_slug_or_name_v1_projects__project_id__milestones_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: slug
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Slug
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Find Milestones By Slug Or Name V1 Projects  Project
                  Id  Milestones Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/milestones:
    post:
      tags:
      - milestones
      summary: Create Milestone
      operationId: create_milestone_v1_milestones_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MilestoneCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MilestoneRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - milestones
      summary: List Milestones
      operationId: list_milestones_v1_milestones_get
      parameters:
      - name: project_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MilestoneRead'
                title: Response List Milestones V1 Milestones Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/milestones/{milestone_id}:
    get:
      tags:
      - milestones
      summary: Get Milestone
      operationId: get_milestone_v1_milestones__milestone_id__get
      parameters:
      - name: milestone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Milestone Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MilestoneRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - milestones
      summary: Update Milestone
      operationId: update_milestone_v1_milestones__milestone_id__patch
      parameters:
      - name: milestone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Milestone Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MilestoneUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MilestoneRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MilestoneCreate:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Slug
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        due_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Due Date
        status:
          type: string
          title: Status
          default: not_started
      type: object
      required:
      - project_id
      - name
      title: MilestoneCreate
    MilestoneRead:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Slug
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        due_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Due Date
        status:
          type: string
          title: Status
          default: not_started
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - project_id
      - name
      - id
      - created_at
      - updated_at
      title: MilestoneRead
    MilestoneUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        due_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Due Date
      type: object
      title: MilestoneUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError